agents/linkedin_agent/requirements.txt
fastapi
uvicorn
pandas
httpx
beautifulsoup4
python-dotenv
